fix: commentList 분리 및 수정
