+++ src/schemas/communities.ts
@@ -60,9 +60,9 @@ export const communityDetailSchema = z.object({
   writer: z.string(),
   writerProfile: z.string().url().nullable(),
   createdAt: z.string(),
-  commentList: z.any(),
+  // commentList: z.any(),
   // commentList: z.array(communityDetailCommentSchema),
-  // commentList: z.array(communityDetailCommentSchema).default([]),
+  commentList: z.array(communityDetailCommentSchema).default([]),
   isLiked: z.boolean(),
   // isPopular: z.booblean(),
   isPopular: z.string().nullable(),
